moved evealuations functions to evaluation module
