+++ data_hub_api/docmaps/provider.py
@@ -12,8 +12,8 @@ from data_hub_api.docmaps.codecs.elife_manuscript import (
     get_docmap_elife_manuscript_output
 )
 from data_hub_api.docmaps.codecs.evaluation import (
-    get_docmap_evaluation_output,
-    get_docmap_evaluation_input
+    iter_docmap_actions_for_evaluations,
+    iter_docmap_evaluation_input,
 )
 from data_hub_api.docmaps.codecs.preprint import (
     get_docmap_preprint_assertion_item,
@@ -25,7 +25,6 @@ from data_hub_api.docmaps.docmap_typing import (
     DocmapAction,
     DocmapAssertion,
     DocmapEvaluationInput,
-    DocmapParticipant,
     DocmapPreprintInput,
     DocmapStep,
     DocmapSteps,
@@ -49,10 +48,6 @@ DOCMAP_ID_PREFIX = (
     +
     'by-publisher/elife/get-by-manuscript-id?'
 )
-
-DOCMAP_EVALUATION_TYPE_FOR_EVALUATION_SUMMARY = 'evaluation-summary'
-DOCMAP_EVALUATION_TYPE_FOR_REPLY = 'reply'
-DOCMAP_EVALUATION_TYPE_FOR_REVIEW_ARTICLE = 'review-article'
 
 ADDITIONAL_MANUSCRIPT_IDS = (
     '80494',
@@ -145,174 +140,13 @@ def get_docmaps_step_for_under_review_status(
     }
 
 
-def get_docmap_assertions_for_preprint_peer_reviewed_step(
+def get_docmap_assertions_for_peer_reviewed_step(
     preprint: dict
 ) -> Sequence[DocmapAssertion]:
     return [{
         'item': get_docmap_preprint_assertion_item(preprint=preprint),
         'status': 'peer-reviewed'
     }]
-
-
-def has_tag_containing(tags: list, text: str) -> bool:
-    return any(
-        text in tag
-        for tag in tags
-    )
-
-
-def get_docmap_evaluation_type_form_tags(
-    tags: list
-) -> Optional[str]:
-    has_author_response_tag = has_tag_containing(tags, 'AuthorResponse')
-    has_summary_tag = has_tag_containing(tags, 'Summary')
-    has_review_tag = has_tag_containing(tags, 'Review')
-    assert not (has_author_response_tag and has_summary_tag)
-    if has_author_response_tag:
-        return DOCMAP_EVALUATION_TYPE_FOR_REPLY
-    if has_summary_tag:
-        return DOCMAP_EVALUATION_TYPE_FOR_EVALUATION_SUMMARY
-    if has_review_tag:
-        return DOCMAP_EVALUATION_TYPE_FOR_REVIEW_ARTICLE
-    return None
-
-
-def get_docmap_evaluation_participants_for_review_article_type() -> DocmapParticipant:
-    return {
-        'actor': {
-            'name': 'anonymous',
-            'type': 'person'
-        },
-        'role': 'peer-reviewer'
-    }
-
-
-def get_related_organization_detail(
-    editor_detail: dict
-) -> str:
-    if editor_detail['country']:
-        return editor_detail['institution'] + ', ' + editor_detail['country']
-    return editor_detail['institution']
-
-
-def get_docmap_evaluation_participants_for_evaluation_summary_type(
-    editor_detail: dict,
-    role: str
-) -> DocmapParticipant:
-    return {
-        'actor': {
-            'name': editor_detail['name'],
-            'type': 'person',
-            '_relatesToOrganization': get_related_organization_detail(editor_detail)
-        },
-        'role': role
-    }
-
-
-def get_docmap_evaluation_participants_for_evalution_summary_type(
-    editor_details_list,
-    senior_editor_details_list
-) -> Sequence[DocmapParticipant]:
-    participants = []
-    for editor_detail in editor_details_list:
-        single_editor_dict = get_docmap_evaluation_participants_for_evaluation_summary_type(
-            editor_detail=editor_detail,
-            role='editor'
-        )
-        participants.append(single_editor_dict)
-    for senior_editor_detail in senior_editor_details_list:
-        single_senior_editor_dict = get_docmap_evaluation_participants_for_evaluation_summary_type(
-            editor_detail=senior_editor_detail,
-            role='senior-editor'
-        )
-        participants.append(single_senior_editor_dict)
-    return cast(Sequence[DocmapParticipant], participants)
-
-
-def get_docmap_evaluation_participants(
-    query_result_item: dict,
-    docmap_evaluation_type: str
-) -> Sequence[DocmapParticipant]:
-    editor_details_list = query_result_item['editor_details']
-    senior_editor_details_list = query_result_item['senior_editor_details']
-    if docmap_evaluation_type == DOCMAP_EVALUATION_TYPE_FOR_REVIEW_ARTICLE:
-        return [get_docmap_evaluation_participants_for_review_article_type()]
-    if docmap_evaluation_type == DOCMAP_EVALUATION_TYPE_FOR_EVALUATION_SUMMARY:
-        return get_docmap_evaluation_participants_for_evalution_summary_type(
-            editor_details_list=editor_details_list,
-            senior_editor_details_list=senior_editor_details_list
-        )
-    return []
-
-
-def get_docmap_actions_for_evaluations(
-    query_result_item: dict,
-    preprint: dict,
-    hypothesis_id: str,
-    evaluation_suffix: str,
-    annotation_created_timestamp: str,
-    docmap_evaluation_type: str
-) -> DocmapAction:
-    return {
-        'participants': get_docmap_evaluation_participants(
-            query_result_item=query_result_item,
-            docmap_evaluation_type=docmap_evaluation_type
-        ),
-        'outputs': [
-            get_docmap_evaluation_output(
-                query_result_item=query_result_item,
-                preprint=preprint,
-                hypothesis_id=hypothesis_id,
-                annotation_created_timestamp=annotation_created_timestamp,
-                evaluation_suffix=evaluation_suffix,
-                docmap_evaluation_type=docmap_evaluation_type
-            )
-        ]
-    }
-
-
-def iter_evaluation_and_type_for_related_preprint_url(
-    evaluations: list,
-    preprint_url: str
-) -> Iterable[Tuple[dict, str]]:
-    for evaluation in evaluations:
-        docmap_evaluation_type = get_docmap_evaluation_type_form_tags(evaluation['tags'])
-        evaluation_preprint_url = evaluation['uri']
-        if evaluation_preprint_url != preprint_url:
-            LOGGER.debug(
-                'ignoring evaluation on another version: %r != %r',
-                evaluation_preprint_url, preprint_url
-            )
-            continue
-        if docmap_evaluation_type in (
-            DOCMAP_EVALUATION_TYPE_FOR_EVALUATION_SUMMARY,
-            DOCMAP_EVALUATION_TYPE_FOR_REVIEW_ARTICLE,
-            DOCMAP_EVALUATION_TYPE_FOR_REPLY
-        ):
-            yield evaluation, docmap_evaluation_type
-
-
-def iter_docmap_actions_for_evaluations(
-    query_result_item: dict,
-    preprint: dict
-) -> Iterable[DocmapAction]:
-    evaluations = query_result_item['evaluations']
-    preprint_url = preprint['preprint_url']
-    for evaluation, docmap_evaluation_type in iter_evaluation_and_type_for_related_preprint_url(
-        evaluations,
-        preprint_url
-    ):
-        hypothesis_id = evaluation['hypothesis_id']
-        annotation_created_timestamp = evaluation['annotation_created_timestamp']
-        evaluation_suffix = evaluation['evaluation_suffix']
-        yield get_docmap_actions_for_evaluations(
-            query_result_item=query_result_item,
-            preprint=preprint,
-            hypothesis_id=hypothesis_id,
-            annotation_created_timestamp=annotation_created_timestamp,
-            evaluation_suffix=evaluation_suffix,
-            docmap_evaluation_type=docmap_evaluation_type
-        )
 
 
 def get_docmaps_step_for_peer_reviewed_status(
@@ -325,30 +159,11 @@ def get_docmaps_step_for_peer_reviewed_status(
             preprint=preprint
             )
         ),
-        'assertions': get_docmap_assertions_for_preprint_peer_reviewed_step(
+        'assertions': get_docmap_assertions_for_peer_reviewed_step(
             preprint=preprint
         ),
         'inputs': [get_docmap_preprint_input(preprint=preprint)]
     }
-
-
-def iter_docmap_evaluation_input(
-    query_result_item: dict,
-    preprint: dict
-):
-    evaluations = query_result_item['evaluations']
-    preprint_url = preprint['preprint_url']
-    for evaluation, docmap_evaluation_type in iter_evaluation_and_type_for_related_preprint_url(
-        evaluations,
-        preprint_url
-    ):
-        evaluation_suffix = evaluation['evaluation_suffix']
-        yield get_docmap_evaluation_input(
-            query_result_item=query_result_item,
-            preprint=preprint,
-            evaluation_suffix=evaluation_suffix,
-            docmap_evaluation_type=docmap_evaluation_type
-        )
 
 
 def get_docmap_inputs_for_revised_steps(
@@ -356,11 +171,13 @@ def get_docmap_inputs_for_revised_steps(
     preprint: dict,
     previous_preprint: dict
 ) -> Sequence[Union[DocmapPreprintInput, DocmapEvaluationInput]]:
-    return list([get_docmap_preprint_input(preprint=preprint)]) + list(
-        iter_docmap_evaluation_input(
+    return (
+        list([get_docmap_preprint_input(preprint=preprint)]) 
+        +
+        list(iter_docmap_evaluation_input(
             query_result_item=query_result_item,
             preprint=previous_preprint
-        )
+        ))
     )
 
 
@@ -381,13 +198,17 @@ def get_docmap_actions_for_revised_steps(
     query_result_item: dict,
     preprint: dict
 ) -> Sequence[DocmapAction]:
-    return list(get_docmap_actions_for_under_review_and_revised_step(
-            query_result_item=query_result_item,
-            preprint=preprint
-        )) + list(iter_docmap_actions_for_evaluations(
+    return (
+        list(get_docmap_actions_for_under_review_and_revised_step(
             query_result_item=query_result_item,
             preprint=preprint
         ))
+        +
+        list(iter_docmap_actions_for_evaluations(
+            query_result_item=query_result_item,
+            preprint=preprint
+        ))
+    )
 
 
 def get_docmaps_step_for_revised_status(
