+++ tests/unit_tests/docmaps/evaluation_test.py
@@ -1,17 +1,28 @@
 from datetime import datetime
+from unittest import mock
+from unittest.mock import patch
 import pytest
-
+from data_hub_api.docmaps.codecs import evaluation as evaluation_module
 from data_hub_api.docmaps.codecs.evaluation import (
+    DOCMAP_EVALUATION_TYPE_FOR_EVALUATION_SUMMARY,
+    DOCMAP_EVALUATION_TYPE_FOR_REPLY,
+    DOCMAP_EVALUATION_TYPE_FOR_REVIEW_ARTICLE,
     DOI_ROOT_URL,
     HYPOTHESIS_URL,
     SCIETY_ARTICLES_ACTIVITY_URL,
     SCIETY_ARTICLES_EVALUATIONS_URL,
+    get_docmap_actions_for_evaluations,
     get_docmap_evaluation_input,
     get_docmap_evaluation_output,
     get_docmap_evaluation_output_content,
     get_docmap_evaluation_output_content_url,
+    get_docmap_evaluation_participants,
+    get_docmap_evaluation_participants_for_evaluation_summary_type,
+    get_docmap_evaluation_participants_for_evalution_summary_type,
+    get_docmap_evaluation_type_form_tags,
     get_elife_evaluation_doi,
-    get_elife_evaluation_doi_url
+    get_elife_evaluation_doi_url,
+    get_related_organization_detail
 )
 
 PREPRINT_DOI_1 = 'preprint_doi_1'
@@ -217,3 +228,145 @@ class TestGetDocmapEvaluationOutput:
                 }
             ]
         }
+
+
+class TestGetEvaluationsTypeFromTags:
+    def test_should_return_evaluation_summary_when_summary_exist_in_tags_list(self):
+        tag_list_with_summary = ['PeerReview', 'evaluationSummary']
+        actual_result = get_docmap_evaluation_type_form_tags(tag_list_with_summary)
+        assert actual_result == DOCMAP_EVALUATION_TYPE_FOR_EVALUATION_SUMMARY
+
+    def test_should_return_review_article_when_review_keyword_exists_in_tags_list(self):
+        tag_list_with_summary = ['PeerReview']
+        actual_result = get_docmap_evaluation_type_form_tags(tag_list_with_summary)
+        assert actual_result == DOCMAP_EVALUATION_TYPE_FOR_REVIEW_ARTICLE
+
+    def test_should_return_review_article_for_review_keyword_even_there_is_undefined_tag(self):
+        tag_list_with_summary = ['PeerReview', 'undefinedTag']
+        actual_result = get_docmap_evaluation_type_form_tags(tag_list_with_summary)
+        assert actual_result == DOCMAP_EVALUATION_TYPE_FOR_REVIEW_ARTICLE
+
+    def test_should_return_reply_when_author_response_keyword_exists_in_tags_list(self):
+        tag_list_with_summary = ['AuthorResponse']
+        actual_result = get_docmap_evaluation_type_form_tags(tag_list_with_summary)
+        assert actual_result == DOCMAP_EVALUATION_TYPE_FOR_REPLY
+
+    def test_should_return_reply_when_author_response_even_there_is_review_tag(self):
+        tag_list_with_summary = ['PeerReview', 'AuthorResponse']
+        actual_result = get_docmap_evaluation_type_form_tags(tag_list_with_summary)
+        assert actual_result == DOCMAP_EVALUATION_TYPE_FOR_REPLY
+
+    def test_should_return_none_when_empty_tags_list(self):
+        tag_list_with_summary = []
+        actual_result = get_docmap_evaluation_type_form_tags(tag_list_with_summary)
+        assert not actual_result
+
+    def test_should_return_none_when_there_is_not_any_defined_tag_in_tags_list(self):
+        tag_list_with_summary = ['undefinedTag']
+        actual_result = get_docmap_evaluation_type_form_tags(tag_list_with_summary)
+        assert not actual_result
+
+    def test_should_raise_error_when_summary_and_author_response_in_tag_list_at_same_time(self):
+        tag_list_with_summary = ['PeerReview', 'evaluationSummary', 'AuthorResponse']
+        with pytest.raises(AssertionError):
+            get_docmap_evaluation_type_form_tags(tag_list_with_summary)
+
+
+class TestGetRelatedOrganizationDetail:
+    def test_should_return_only_institution_if_country_not_defined(self):
+        editor_detail_dict = {'institution': 'institution_1', 'country': None}
+        result = get_related_organization_detail(editor_detail_dict)
+        assert result == 'institution_1'
+
+    def test_should_return_institution_and_country_if_defined(self):
+        editor_detail_dict = {'institution': 'institution_1', 'country': 'country_1'}
+        result = get_related_organization_detail(editor_detail_dict)
+        assert result == 'institution_1, country_1'
+
+class TestGetDocmapEvaluationParticipantsForEvaluationSummaryType:
+    def test_should_raise_assertion_error_if_the_role_not_editor_or_senior_editor(self):
+        editor_detail_dict = {'name': 'name_1', 'institution': 'institution_1', 'country': None}
+        role = 'role_1'
+        with pytest.raises(AssertionError):
+            get_docmap_evaluation_participants_for_evaluation_summary_type(
+                editor_detail=editor_detail_dict,
+                role=role
+            )
+
+    def test_should_populate_participants_with_and_given_reviewing_editor_detail_(self):
+        editor_detail_dict = {'name': 'name_1', 'institution': 'institution_1', 'country': None}
+        role = 'editor'
+        result = get_docmap_evaluation_participants_for_evaluation_summary_type(
+            editor_detail=editor_detail_dict,
+            role=role
+        )
+        assert result == {
+            'actor': {
+                'name': 'name_1',
+                'type': 'person',
+                '_relatesToOrganization': get_related_organization_detail(editor_detail_dict)
+            },
+            'role': 'editor'
+        }
+
+
+class TestGetDocmapEvaluationParticipantsForEvalutionSummaryType:
+    def test_should_populate_evaluation_participants_for_evaluation_summary_type(self):
+        editor_detail_1 = {
+            'name': 'editor_name_1',
+            'institution': 'editor_institution_1',
+            'country': 'editor_country_1'
+        }
+        editor_detail_2 = {
+            'name': 'editor_name_2',
+            'institution': 'editor_institution_2',
+            'country': None
+        }
+        senior_editor_detail_1 = {
+            'name': 'senior_editor_name_1',
+            'institution': 'senior_editor_institution_1',
+            'country': 'senior_editor_country_1'
+        }
+        result = get_docmap_evaluation_participants_for_evalution_summary_type(
+            editor_details_list=[editor_detail_1, editor_detail_2],
+            senior_editor_details_list=[senior_editor_detail_1]
+        )
+        assert result == [
+            get_docmap_evaluation_participants_for_evaluation_summary_type(
+                editor_detail=editor_detail_1,
+                role= 'editor'
+            ),
+            get_docmap_evaluation_participants_for_evaluation_summary_type(
+                editor_detail=editor_detail_2,
+                role= 'editor'
+            ),
+            get_docmap_evaluation_participants_for_evaluation_summary_type(
+                editor_detail=senior_editor_detail_1,
+                role= 'senior-editor'
+            )
+        ]
+
+
+class TestGetDocmapEvaluationParticipants:
+    def test_should_call_func_review_article_type_when_evaluation_type_review_article(self):
+        with patch.object(
+            evaluation_module,
+            'get_docmap_evaluation_participants_for_review_article_type'
+        ) as mock:
+            get_docmap_evaluation_participants(
+                query_result_item=DOCMAPS_QUERY_RESULT_ITEM_1,
+                docmap_evaluation_type=DOCMAP_EVALUATION_TYPE_FOR_REVIEW_ARTICLE
+            )
+            mock.assert_called_once()
+
+    def test_should_call_func_evaluation_summary_type_when_evaluation_type_valuation_summary(self):
+        with patch.object(
+            evaluation_module,
+            'get_docmap_evaluation_participants_for_evalution_summary_type'
+        ) as mock:
+            get_docmap_evaluation_participants(
+                query_result_item=DOCMAPS_QUERY_RESULT_ITEM_1,
+                docmap_evaluation_type=DOCMAP_EVALUATION_TYPE_FOR_EVALUATION_SUMMARY
+            )
+            mock.assert_called_once()
+
